Don't use ES product collections when indexing
I was getting empty product collections on anchor categories while the index was running. 

P.S. I'm actually using the paid version, so just copied the changes I made to my version. 
It's a great base to start some heavy customisations on, except that having no events to listen on (and the code being in the local pool) make it almost necessary to hack at some of the files. I'd love to add some more events into the code where it would allow me to do things like modify the query.